+++ tests/conftest.py
@@ -1,5 +1,3 @@
-import os
-from glob import glob
 from pathlib import Path
 from typing import Dict, List, Union
 from unittest.mock import patch
