+++ tests/test_make_files.py
@@ -1,5 +1,3 @@
-from pathlib import Path
-
 from conftest import make_files, read_files
